Changes made.

+++ MauiAssignment7/MainPage.xaml.cs
@@ -4,6 +4,8 @@ namespace MauiAssignment7
 {
     public partial class MainPage : ContentPage
     {
+        public string fullFileName = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "myFile.txt");
+
         private User _user;
         public User CurrentUser
         {
@@ -21,16 +23,20 @@ namespace MauiAssignment7
 
         public void SaveData(User user)
         {
-            string filePath = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
-            string fullFileName = Path.Combine(filePath, "myFile.txt");
-            string userJson = JsonConvert.SerializeObject(user);
-            File.WriteAllText(fullFileName, userJson);
+            //string filePath = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
+            //string fullFileName = Path.Combine(filePath, "myFile.txt");
+            if (File.Exists(fullFileName))
+            {
+                string userJson = JsonConvert.SerializeObject(user);
+                File.WriteAllText(fullFileName, userJson);
+            }
+            else new User();
         }
 
         public User LoadData()
         {
-            string filePath = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
-            string fullFileName = Path.Combine(filePath, "myFile.txt");
+            //string filePath = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
+            //string fullFileName = Path.Combine(filePath, "myFile.txt");
             if (File.Exists(fullFileName))
             {
                 string fileContent = File.ReadAllText(fullFileName);
@@ -42,7 +48,17 @@ namespace MauiAssignment7
 
         private void Button_Clicked(object sender, EventArgs e)
         {
+            //string filePath = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
+            //string fullFileName = Path.Combine(filePath, "myFile.txt");
             SaveData(CurrentUser);
+            if (File.Exists(fullFileName))
+            {
+                DisplayAlert("Alert", "User account saved.", "OK");
+            }
+            else 
+            {
+                DisplayAlert("Alert", "User account not saved.", "OK");
+            }
         }
     }
 }
